requirements.txt
flask
playwright
requests
beautifulsoup4
pandas
python-dotenv
pdfplumber
